+++ projects/ontimize-web-ngx/src/lib/util/codes.ts
@@ -111,6 +111,9 @@ export class Codes {
   public static OAppLayoutMode = ['mobile', 'desktop'];
   public static OSidenavMode = ['over', 'push', 'side'];
 
+  public static APP_LAYOUT_MODE_DESKTOP: OAppLayoutMode = 'desktop';
+  public static APP_LAYOUT_MODE_MOBILE: OAppLayoutMode = 'mobile';
+
   static isDoubleClickMode(value: string): boolean {
     return Codes.DETAIL_MODE_DBLCLICK_VALUES.indexOf(value) !== -1;
   }